requirements.txt
Twisted[tls]
service-identity
ldaptor
PyYAML
